% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ct_get}
\alias{ct_get}
\title{Retrieve posts from CrowdTangle API}
\usage{
ct_get(token, ids, sort_by = "date", count = 100, from, to)
}
\arguments{
\item{token}{Credentials/token.}

\item{ids}{Id of saved post lists of the dashboard associated with the token.}

\item{sort_by}{Defaults to "date".}

\item{count}{An integer value, max 100.}

\item{from}{Begin date, format YYYY-MM-DD.}

\item{to}{End date, format YYYY-MM-DD.}
}
\description{
Retrieve posts from CrowdTangle API
}
\examples{
token <- readLines("~/.crowdtangle/token")
y <- ct_get(
  token = token,
  ids = 1771322,
  from = "2021-01-01",
  to = "2021-01-31"
)
}
