% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ct_lists}
\alias{ct_lists}
\title{Get lists saved at Dashboard.}
\usage{
ct_lists(token)
}
\arguments{
\item{token}{The CrowdTangle token.}
}
\description{
Get lists saved at Dashboard.
}
\examples{
token <- readLines("~/.crowdtangle/token")
ct_lists(token)
}
