% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-package.R
\docType{package}
\name{fb-package}
\alias{fb-package}
\title{The fb R Package}
\description{
Functionality for retrieving and processing Facebook data.
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
