% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager.R
\docType{data}
\name{fp_colnames}
\alias{fp_colnames}
\title{Column names of Facepager output}
\format{
An object of class \code{character} of length 18.
}
\usage{
fp_colnames
}
\description{
Column names of Facepager output
}
\keyword{datasets}
