% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_report.R
\name{ct_read_csv}
\alias{ct_read_csv}
\alias{historical_report_purge}
\title{Read historical report}
\usage{
ct_read_csv(x, verbose = TRUE, ...)

historical_report_purge(data, verbose = TRUE)
}
\arguments{
\item{x}{Path of file to read. Expected to be historical report from
Crowdtangle.}

\item{verbose}{A \code{logical} value, whether to output progress messages.}

\item{...}{Arguments passed into \code{aws.s3::get_object()}, if data is retrieved
from S3.}

\item{data}{A \code{tibble} with parsed historical report.}
}
\description{
Read historical report
}
\details{
\code{historical_report_purge()} drops posts with \code{NA} content. These
may posts with picture content only.
}
\examples{
fname <- "~/Downloads/2023-05-16-04-54-58-EDT-Historical-Report-PartyPages2-2014-01-01--2015-01-01.csv"

tbl <- ct_read_csv(fname)
tbl <- historical_report_purge(tbl)
}
