% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partypages-package.R
\docType{package}
\name{partypages-package}
\alias{partypages-package}
\title{The PartyPages R Package}
\description{
Data and supplementary material for the Party Pages Dataset.
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
