% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partypages-package.R
\docType{data}
\name{partypages}
\alias{partypages}
\title{Metadata for posts}
\format{
\subsection{\code{partypages}}{

\describe{
\item{User Name}{}
\item{user_name}{Identical with column 'User Name'. Better naming convention.}
\item{Page Name}{}
\item{country}{}
\item{Facebook Id}{A \code{double} vector, mixed up.}
\item{facebook_id}{A \code{character} vector, preliminary consolidation.}
\item{Likes at Posting}{}
\item{Followers at Posting}{}
\item{Created}{}
\item{date}{}
\item{Type}{}
\item{Likes}{}
\item{Comments}{}
\item{Shares}{}
\item{URL}{}
\item{Total Interactions}{}
}
}
}
\usage{
partypages
}
\description{
Metadata for Facebook pages covered by the corpus for posts from 2014-12-31
to 2020-12-30. Note that the coverage of the corpus goes beyond the metadata
table: It includes posts from 2014-07-12 to end of 2014 and 2021-01-01.
}
\keyword{datasets}
