% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partypages-package.R
\docType{data}
\name{partypages2}
\alias{partypages2}
\title{Metadata for posts (updated)}
\format{
\subsection{\code{partypages}}{

\describe{
\item{Page Name}{}
\item{User Name}{}
\item{Facebook Id}{}
\item{Page Category}{}
\item{Page Admin Top Country}{}
\item{Page Description}{}
\item{Page Created}{}
\item{Likes at Posting}{}
\item{Followers at Posting}{}
\item{Post Created}{}
\item{Post Created Date}{}
\item{Post Created Time}{}
\item{Type}{}
\item{Total Interactions}{}
\item{Likes}{}
\item{Comments}{}
\item{Shares}{}
\item{Love}{}
\item{Wow}{}
\item{Haha}{}
\item{Sad}{}
\item{Angry}{}
\item{Care}{}
\item{Video Share Status}{}
\item{Is Video Owner?}{}
\item{Post Views}{}
\item{Total Views}{}
\item{Total Views For All Crossposts}{}
\item{Video Length}{}
\item{URL}{}
\item{Link}{}
\item{Final Link}{}
\item{Image Text}{}
\item{Link Text}{}
\item{Description}{}
\item{Sponsor Id}{}
\item{Sponsor Name}{}
\item{Sponsor Category}{}
\item{Overperforming Score (weighted  —  Likes 1x Shares 1x Comments 1x Love 1x Wow 1x Haha 1x Sad 1x Angry 1x Care 1x )}{}
\item{date}{}
}
}
}
\usage{
partypages2
}
\description{
Metadata for Facebook pages covered by the corpus for posts from 2014-01-01
to 2023-05-16. Supplements corpus PARTYPAGES v0.1.0.
}
\keyword{datasets}
