% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partypages-package.R
\docType{data}
\name{partypages_mapping}
\alias{partypages_mapping}
\title{Mapping table for PartyPages dataset.}
\format{
\subsection{\code{partypages_mapping}}{

A tibble with ... rows and 2 columns:
\describe{
\item{user_name}{A \code{character} vector with user names.}
\item{facebook_id}{A \code{character} vector.}
}
}
}
\usage{
partypages_mapping
}
\description{
A tibble for mapping User Names, Facebooks Id and more.
}
\details{
The column 'facebook_id' is technically an integer number. But values exceed
the 10^9 limit for R integer vectors(see ?integer), so IDs are represented
using a character vector for technical reasons.
}
\keyword{datasets}
